+++ onchain/rollups/contracts/dapp/CartesiDAppFactory.sol
@@ -3,6 +3,7 @@
 
 pragma solidity ^0.8.8;
 
+import {ENS} from "@ensdomains/ens-contracts/contracts/registry/ENS.sol";
 import {Create2} from "@openzeppelin/contracts/utils/Create2.sol";
 
 import {ICartesiDAppFactory} from "./ICartesiDAppFactory.sol";
@@ -12,6 +13,12 @@ import {CartesiDApp} from "./CartesiDApp.sol";
 /// @title Cartesi DApp Factory
 /// @notice Allows anyone to reliably deploy a new `CartesiDApp` contract.
 contract CartesiDAppFactory is ICartesiDAppFactory {
+    ENS public immutable ens;
+
+    constructor(ENS _ens) {
+        ens = _ens;
+    }
+
     function newApplication(
         IConsensus _consensus,
         address _dappOwner,
@@ -19,6 +26,7 @@ contract CartesiDAppFactory is ICartesiDAppFactory {
     ) external override returns (CartesiDApp) {
         CartesiDApp application = new CartesiDApp(
             _consensus,
+            ens,
             _dappOwner,
             _templateHash
         );
@@ -41,6 +49,7 @@ contract CartesiDAppFactory is ICartesiDAppFactory {
     ) external override returns (CartesiDApp) {
         CartesiDApp application = new CartesiDApp{salt: _salt}(
             _consensus,
+            ens,
             _dappOwner,
             _templateHash
         );
@@ -67,7 +76,7 @@ contract CartesiDAppFactory is ICartesiDAppFactory {
                 keccak256(
                     abi.encodePacked(
                         type(CartesiDApp).creationCode,
-                        abi.encode(_consensus, _dappOwner, _templateHash)
+                        abi.encode(_consensus, ens, _dappOwner, _templateHash)
                     )
                 )
             );